Add dummy header links
Inspired by boston.rb's website

Add listing of major links in the header. URL is not yet included (since
many of those pages don't yet exist).

The style is pretty simple, but the links grow smaller on a medium
screen and become more mobile-accessible (large and spaced out) on a
small screen.